# go-backend/docs/swagger.yaml
basePath: /api
definitions:
  models.Category:
    properties:
      catID:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  models.MediaAsset:
    properties:
      altText:
        type: string
      createdAt:
        type: string
      id:
        type: string
      postId:
        type: string
      type:
        type: string
      uploadedBy:
        type: string
      url:
        type: string
    type: object
  models.Post:
    properties:
      authorId:
        type: string
      categoryId:
        type: string
      content:
        type: string
      coverImage:
        $ref: '#/definitions/models.MediaAsset'
      createdAt:
        type: string
      date:
        type: string
      gallery:
        items:
          $ref: '#/definitions/models.MediaAsset'
        type: array
      id:
        type: string
      summary:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      type:
        description: article, news, tutorial
        type: string
      updatedAt:
        type: string
    type: object
  models.Tag:
    properties:
      id:
        type: string
      name:
        type: string
      tagID:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend APIs for Yahya's RemixJS Blog
  title: YahyaOnCloud Blog API
  version: "1.0"
paths:
  /categories:
    delete:
      description: Deletes a category by ID
      parameters:
      - description: Category ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a category
      tags:
      - categories
    get:
      description: Returns a list of all blog categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Accepts a single category object or an array of category objects
        for bulk insert
      parameters:
      - description: List of categories or single category
        in: body
        name: categories
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Category'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create category or list of categories
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Updates an existing blog category
      parameters:
      - description: Category ID
        in: query
        name: id
        required: true
        type: string
      - description: Updated category object
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a category
      tags:
      - categories
  /posts:
    get:
      description: Returns a list of all blog posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Accepts a post object and creates it
      parameters:
      - description: Post to create
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Deletes a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete a post
      tags:
      - posts
    get:
      description: Returns a single post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Accepts a post object and updates it
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post to update
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Update a post
      tags:
      - posts
  /tags:
    delete:
      description: Deletes a tag by ID
      parameters:
      - description: Tag ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a tag
      tags:
      - tags
    get:
      description: Returns a list of all blog tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tag'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Accepts a single tag object or an array of tag objects for bulk
        insert
      parameters:
      - description: List of tags or single tag
        in: body
        name: tags
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Tag'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create tag or list of tags
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Updates an existing blog tag
      parameters:
      - description: Tag ID
        in: query
        name: id
        required: true
        type: string
      - description: Updated tag object
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/models.Tag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tag'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a tag
      tags:
      - tags
swagger: "2.0"