# go-backend/models/categories.go
package models

import "go.mongodb.org/mongo-driver/bson/primitive"

type Category struct {
    ID   primitive.ObjectID `bson:"_id,omitempty" json:"id,omitempty"`
	CatID string             `bson:"catID" json:"catID"`
    Name string             `bson:"name" json:"name"`
    Slug string             `bson:"slug" json:"slug"`
}